player rotation has been added

+++ src/handle_keys_click.c
@@ -4,22 +4,22 @@ void update_direction(t_game *game)
 {
   // printf("x ---> %f\n", game->n_px); 
   // printf("y ---> %f\n", game->n_py);
-  float dx;
-  float dy;
+  double dx;
+  double dy;
 
   dx = game->n_px - game->player_x;
   dy = game->n_py - game->player_y;
 
-  float steps;
+  double steps;
   if(fabs(dx) > fabs(dy))
     steps = fabs(dx);
   else
     steps = fabs(dy);
   int i;
-  float x = game->player_x;
-  float y = game->player_y;
-  float x_inc = dx / steps;
-  float y_inc = dy / steps;
+  double x = game->player_x;
+  double y = game->player_y;
+  double x_inc = dx / steps;
+  double y_inc = dy / steps;
   i = 0;
   while(i < steps)
   {
